Separate analytics area widgets to their own files
